issue-23: added symlinks for XCode
This should not have been necessary, but I couldn't figure out
how to get XCode to look in the directory containing the source
files for local headers.

This commit also contains edits to configure.ac to get the system
to build on the latest Mac.  Not that this is still imperfect, linking
fails with glXSwapBuffers undefined, but I could make that go away by
adding -lGL manually at the end, which may be getting the library from
homebrew in /opt/homebrew/lib...  I played around with installing XQuartz
homebrew but that was problematic.  I'm really not sure what is the current
state of the machine (nutmeg), but it is working, which has taken days!?